HYP-133: removing unique table elements away from table component.
