+++ src/Table/index.tsx
@@ -1,8 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
 
-import { ClockIcon, PadlockIcon } from '../index.js'
-
 export interface TableProps {
   styles?: {
     tr: React.CSSProperties
@@ -14,34 +12,6 @@ export interface TableProps {
 }
 
 type RowValue = React.ReactNode | number | string | null | undefined
-
-export const NameCell = ({ name, date }: { name: string; date: string }) => (
-  <Col>
-    <Row>
-      <PadlockIcon />
-      <Title>{name}</Title>
-    </Row>
-    <Row>
-      <ClockIcon />
-      <Date>{date}</Date>
-    </Row>
-  </Col>
-)
-
-export const StatusCell = ({
-  color = '#27847a',
-  Icon,
-  status,
-}: {
-  color?: string
-  Icon: React.FC
-  status: string
-}) => (
-  <Col>
-    <Icon />
-    <Status color={color}>{status}</Status>
-  </Col>
-)
 
 const Table: React.FC<React.PropsWithChildren<TableProps>> = (props) => {
   return (
@@ -69,53 +39,6 @@ const Table: React.FC<React.PropsWithChildren<TableProps>> = (props) => {
     </Background>
   )
 }
-
-const Title = styled('div')`
-  color: #6aa685;
-  font-family: Inter;
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 0px; /* 0% */
-`
-
-const Status = styled('div')`
-  width: 90px;
-  background-color: ${(props) => props.color || '#27847a'};
-  color: #fff;
-  border-radius: 60px;
-
-  text-align: center;
-  font-family: Roboto;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 10px; /* 83.333% */
-`
-
-const Date = styled('div')`
-  color: #27847a;
-
-  font-family: Inter;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 10px; /* 83.333% */
-`
-
-const Row = styled('div')`
-  display: flex;
-  flex-direction: row;
-  gap: 5px;
-  align-items: center;
-`
-
-const Col = styled('div')`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-`
 
 const TH = styled('th')`
   border-bottom: solid 1px #fff;
@@ -155,10 +78,9 @@ const TR = styled('tr')`
     ${(props) =>
       props.color !== 'not-page' &&
       `
-    border-bottom: 10px solid #eee;
-    border-left: 10px solid #27847a`};
-
-    box-shadow: 1px 1px 1px rgba(0, 0, 0, 0.3);
+    right: 30px;
+    border-bottom: 15px solid #eee;
+    border-right: 15px solid #27847a`};
   }
 `
